Refactor to simplify file handling
