+++ .env.default
@@ -65,10 +65,6 @@ CDN__BUCKET_ANSWER=media
 CDN__BUCKET=media
 CDN__BUCKET_OPERATIONS=media
 CDN__STORAGE_ADDRESS=http://localhost:9000
-CDN__LEGACY_REGION=
-CDN__LEGACY_BUCKET=
-CDN__LEGACY_SECRET_KEY=
-CDN__LEGACY_ACCESS_KEY=
 CDN__TTL_SIGNED_URLS=3600
 
 # jsonld converter
