+++ pyproject.toml
@@ -55,9 +55,7 @@ env = [
     "DATABASE__USER=postgres",
     "DATABASE__DB=test",
     "ARBITRARY_DB=test_arbitrary",
-    "TASK_ANSWER_ENCRYPTION__BATCH_LIMIT=1",
-    "CDN__LEGACY_REGION=us-east-1",
-    "CDN__LEGACY_BUCKET=testing"
+    "TASK_ANSWER_ENCRYPTION__BATCH_LIMIT=1"
 ]
 
 [tool.coverage.run]
