+++ src/apps/applets/commands/applet_ema.py
@@ -34,7 +34,7 @@ from apps.workspaces.domain.constants import Role
 from config import settings
 from infrastructure.commands.utils import coro
 from infrastructure.database import atomic, session_manager
-from infrastructure.dependency.cdn import get_operations_bucket
+from infrastructure.storage.buckets import get_operations_bucket
 from infrastructure.utility import CDNClient, ObjectNotFoundError
 
 app = typer.Typer()
