+++ src/infrastructure/storage/cdn_client.py
@@ -8,14 +8,13 @@ from typing import BinaryIO
 
 import boto3
 import httpx
-from botocore import UNSIGNED
 from botocore.config import Config
 from botocore.exceptions import ClientError, EndpointConnectionError
 
 from apps.file.errors import FileNotFoundError
 from apps.shared.exception import NotFoundError
 from infrastructure.logger import logger
-from infrastructure.utility.cdn_config import CdnConfig
+from infrastructure.storage.cdn_config import CdnConfig
 
 
 class ObjectNotFoundError(Exception):
@@ -32,7 +31,7 @@ class CDNClient:
     def __init__(self, config: CdnConfig, env: str, *, max_concurrent_tasks: int = 10):
         self.config = config
         self.env = env
-        self.client = self.configure_client(config)
+        self.client = self._configure_client(config)
 
         # semaphore for concurrent calls of urlib3 in boto3
         self.semaphore = asyncio.Semaphore(max_concurrent_tasks)
@@ -46,11 +45,10 @@ class CDNClient:
     def generate_private_url(self, key):
         return f"s3://{self.config.bucket}/{key}"
 
-    def configure_client(self, config, signature_version=None):
+    def _configure_client(self, config):
         assert config, "set CDN"
         client_config = Config(
             max_pool_connections=25,
-            signature_version=signature_version,
         )
 
         if config.access_key and config.secret_key:
@@ -119,17 +117,6 @@ class CDNClient:
         media_type = mimetypes.guess_type(key)[0] if mimetypes.guess_type(key)[0] else "application/octet-stream"
         return file, media_type
 
-    def _generate_public_url(self, key):
-        client = self.configure_client(config=self.config, signature_version=UNSIGNED)
-        url = client.generate_presigned_url(
-            "get_object",
-            Params={
-                "Bucket": self.config.bucket,
-                "Key": key,
-            },
-            ExpiresIn=0,
-        )
-        return url
 
     def _generate_presigned_url(self, key):
         url = self.client.generate_presigned_url(
@@ -148,11 +135,6 @@ class CDNClient:
             url = await asyncio.wrap_future(future)
             return url
 
-    async def generate_public_url(self, key):
-        with ThreadPoolExecutor() as executor:
-            future = executor.submit(self._generate_public_url, key)
-            url = await asyncio.wrap_future(future)
-            return url
 
     async def delete_object(self, key: str | None):
         async with self.semaphore:
@@ -167,9 +149,11 @@ class CDNClient:
                 result = await asyncio.wrap_future(future)
                 return result.get("Contents", [])
 
-    def generate_presigned_post(self, bucket, key):
+
+
+    def generate_presigned_post(self, key):
         # Not needed ThreadPoolExecutor because there is no any IO operation (no API calls to s3)
-        return self.client.generate_presigned_post(bucket, key, ExpiresIn=self.config.ttl_signed_urls)
+        return self.client.generate_presigned_post(self.config.bucket, key, ExpiresIn=self.config.ttl_signed_urls)
 
     def _copy(self, key, storage_from: "CDNClient", key_from: str | None = None) -> int:
         key_from = key_from or key
