+++ src/infrastructure/storage/cdn_config.py
