+++ src/infrastructure/utility/__init__.py
@@ -1,3 +1,3 @@
-from infrastructure.utility.cdn_client import *  # noqa: F401, F403
+from infrastructure.storage.cdn_client import *  # noqa: F401, F403
 from infrastructure.utility.notification_client import *  # noqa: F401, F403
 from infrastructure.utility.redis_client import *  # noqa: F401, F403
